Work around a header name collission on util.h
Perl provides a util.h.
openpty() comes from a system util.h in OSX.

The end result is that the system util.h is unreachable,
and so openpty() is never declared.

This commit works around this in a nasty and non-portable
fashiom.  First, it adds a second way of including the header
in the XS code:

    #ifdef UTIL_H_ABS_PATH
    # define UTIL_H_ABS_PATH
    #endif

And second, it uses an arcane command to figure out
the exact header that gcc or clang ended up finding.

It's pretty terrible, all things told, but since it'll only
really kick in for OSX, the horror is somewhat contained.